requirements.txt
# Core (keine schweren Abhängigkeiten bewusst minimal)
python-dotenv>=1.0.0
pytest>=8.2.0
fastapi>=0.111.0
uvicorn>=0.30.0
